chore(refactor): understandable directory structure
