+++ Backend/server.js
@@ -3,12 +3,12 @@ const express = require("express");
 const app = express();
 
 // Setup all middleware
-const setupMiddleware = require("./middleware");
+const setupMiddleware = require("./api/middleware/index.middleware.js");
 setupMiddleware(app);
 
 // Routes
-const products = require("./api/product");
-const home = require("./api/home");
+const products = require("./api/routes/product.routes.js");
+const home = require("./api/routes/home.routes.js");
 
 app.use("/products", products);
 app.use("/", home);
